+++ web/modules/custom/ai_screening/src/Helper/ThemeHelper.php
@@ -34,7 +34,7 @@ final class ThemeHelper extends AbstractHelper implements EventSubscriberInterfa
     $form = &$event->getForm();
     $formId = $event->getFormId();
     if ('node_project_form' === $formId) {
-      $form['#title'] = t('Create new project');
+      $form['#title'] = t('Create new screening');
     }
     if ('node_static_form' === $formId) {
       $form['#title'] = t('Create new static page');